Added a basic abstract class Component. It has a queuing mechanism for asynchronously receiving messages from several Receivers. Subclasses need to implement processMessage(SEMAINEMessage).
The asynchronous notification mechanism for Receivers was adapted so that listeners will be alerted 
of the fact that a message is available, but they can decide from which thread and when to pick up
the message. The receiver thread will block until the message was picked up, which for slow clients
has the advantage that the JMS server will look after discarding outdated messages, rather than them piling up in our processing queue. 

git-svn-id: https://semaine.opendfki.de/repos/trunk@5 2252c265-7fd8-49a5-b10e-6a9b7765d42e