Simplification of StateInfo class structure  in view of C++ port. Functionality unchanged, only syntax changes.
Removed classes AgentStateReceiver, DialogStateReceiver and UserStateReceiver, as well as the corresponding SEMAINEAgentStateMessage, SEMAINEDialogStateMessage and SEMAINEUserStateMessage. These were really only type-specific front-ends for StateReceiver and SEMAINEStateMessage, respectively.

Now all StateInfo objects have a method getType() which states whether it is an AgentStateInfo, DialogStateInfo, UserStateInfo or (new) ContextStateInfo object. The revised implementation of StateReceiver knows the Type of StateInfo objects it should create (this information is provided by the component writer in the constructor).

All code in SEMAINE/trunk/java/src has been syntactically updated to the new format (but not tested).

The ContextStateInfo is not yet operable, it will be when the StateInfo parsing is updated to XPath.

git-svn-id: https://semaine.opendfki.de/repos/trunk@304 2252c265-7fd8-49a5-b10e-6a9b7765d42e