New state info implementation based on XPath. This breaks compatibility with previous users of 'current best guess' components. Watch error messages to see which info keys are problematic.
This implementation uses a shortname-to-XPath config file which follows the wiki as of 23-07-2009. Without changes in the source code, new information can be added and encoded/decoded in namespace-aware XML.

This implementation uses a built-in config file (see src/eu/semaine/datatypes/stateinfo/stateinfo.config). In order to override this with a modified config file, start the java process with system property -Dconfig.stateinfo=/path/to/my-stateinfo.config

git-svn-id: https://semaine.opendfki.de/repos/trunk@313 2252c265-7fd8-49a5-b10e-6a9b7765d42e