BREAKING CHANGE! Code using EmotionML must be changed so it compiles again.
Updating EmotionML.java to be in line with the Second Public Working Draft of EmotionML, http://www.w3.org/TR/2010/WD-emotionml-20100729/

This version removes support for old Elements such as <dimensions> and old Attributes such as "set", and adds a complete list of element and attribute names from the latest working draft.

The file also contains string constants for a number of Emotion vocabularies relevant in SEMAINE, and convenience functions for converting value ranges between SEMAINE dimensions and FSRE dimensions.

git-svn-id: https://semaine.opendfki.de/repos/trunk@563 2252c265-7fd8-49a5-b10e-6a9b7765d42e