Updated speech code so that it can use one of several voices proposed for a given character in character-config.xml.
Concretely, we want to use the unit selection voices if they are installed, and fall back to the HMM voices otherwise. This gives us the possibility to ship SEMAINE with only HMM voices (which is small and versatile) and let the user install the unit selection voices if they want better quality.

git-svn-id: https://semaine.opendfki.de/repos/trunk@573 2252c265-7fd8-49a5-b10e-6a9b7765d42e