I believe this fixes ticket:50.
The problem was more precisely that the UtteranceActionProposer did not propose any utterances. This occurred only when the system was started using the jar files; things worked OK when the system was started e.g. from within Eclipse.

The problem was that the template files were not included in semaine-dialogue.jar -- this was easy to fix.

git-svn-id: https://semaine.opendfki.de/repos/trunk@580 2252c265-7fd8-49a5-b10e-6a9b7765d42e