A first version of the new java-based EmotionFusion component. It seems to basically work but needs thorough testing.
This component listens to all subtopics of semaine.data.state.user.emma.emotion.*, and computes consolidated statements about user emotion by computing weighted averages of the values, where the "confidence" is used as the weight.

This works for all dimensions, appraisals, action-tendencies, categories, and for intensity. For category, the resulting confidence is the average of the incoming confidences; for the dimensions etc., the resulting confidence is one minus the standard deviation of the value across data points (variance being computed using confidence as the weight).

Two config settings (in the java config file used to start the system) can be used to influence the behaviour of the EmotionFusion:

semaine.emotion-fusion.max-delay  -- specifies the maximum waiting time, in milliseconds, before the fused data is sent. Default: 500 ms
semaine.emotion-fusion.num-raters -- can be used to indicate a number of "expected" data points per emotion annotation; as soon as this number is reached, the fused data is sent. Default: 999, i.e. this will not be reached unless explicitly set by a user.


This checkin also contains a first version of the Random emotion generator, needed for testing. It will send random values for the four FSRE dimensions, intensity, and random confidences for the three interest categories. Its average sending period can be influenced using the config setting:

semaine.random-emotion.period -- the average time delay (in milliseconds) between one emotion message and the next. Default: 1000 ms 

git-svn-id: https://semaine.opendfki.de/repos/trunk@582 2252c265-7fd8-49a5-b10e-6a9b7765d42e