Created a simple state logger which writes all state updates (user, agent, dialog and context state) into a csv file. The columns correspond to the entries in stateinfo.config, including the order in which they are shown.
In order to have the agent utterances as well in the log, I created an AgentBehaviourObserver which enters into the agent state the new fields "agentUtterance" and "agentUtteranceStartTime", which are filled with speaker and listener utterances in both direct and prepare branches. The time is the actual start time as reported by the player callback.

git-svn-id: https://semaine.opendfki.de/repos/trunk@689 2252c265-7fd8-49a5-b10e-6a9b7765d42e