Updating code to use ActiveMQ-CPP 3.1.3.
THIS IS A BREAKING CHANGE, so update only if you need this for now. Tested on Mac OS X for now; will test on linux and Windows later this week.

The added benefit is that we get proper error messages when things fail in the c++-to-ActiveMQ communication.

The breaking change is that, in order for this to run, you must add three lines into your main.cpp:

#include <activemq/library/ActiveMQCPP.h>
[...]
activemq::library::ActiveMQCPP::initializeLibrary();
[...]
activemq::library::ActiveMQCPP::shutdownLibrary();

See for example c++/src/semaine/executables/jimandjohn.cpp how to add these.

git-svn-id: https://semaine.opendfki.de/repos/trunk@769 2252c265-7fd8-49a5-b10e-6a9b7765d42e